new test_selector.py and refactor test file names due to the collision
